move function to common JS
