+++ blocks/eloqua-form/eloqua-form.js
@@ -1,6 +1,5 @@
 import { hideSidebar } from '../../common/sidebar/sidebar.js';
-import { isEloquaFormAllowed } from '../../scripts/scripts.js';
-import { getTextLabel } from '../../scripts/common.js';
+import { getTextLabel, isEloquaFormAllowed } from '../../scripts/common.js';
 
 // eslint-disable no-console
 const addForm = async (block) => {
