+++ scripts/scripts.js
@@ -51,19 +51,6 @@ function isCTALinkCheck(ctaLink) {
 }
 
 /**
- * Check if one trust group is checked.
- * @param {String} groupName the one trust croup like: C0002
- */
-export function checkOneTruckGroup(groupName) {
-  const oneTrustCookie = decodeURIComponent(document.cookie.split(';').find((cookie) => cookie.trim().startsWith('OptanonConsent=')));
-  return oneTrustCookie.includes(`${groupName}:1`);
-}
-
-export function isEloquaFormAllowed() {
-  return checkOneTruckGroup('C0004');
-}
-
-/**
  * Returns a picture element with webp and fallbacks / allow multiple src paths for every breakpoint
  * @param {string} src Default image URL (if no src is passed to breakpoints object)
  * @param {boolean} eager load image eager
